mediacapture-streams/MediaStream-gettrackid.https.html

<!doctype html>
<title>Retrieving a track from a MediaStream</title>
<link rel="author" title="Dominique Hazael-Massieux" href="mailto:dom@w3.org"/>
<link rel="help" href="http://dev.w3.org/2011/webrtc/editor/getusermedia.html#widl-MediaStream-getTrackById-MediaStreamTrack-DOMString-trackId">
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>

<p class="instructions">When prompted, accept to share your video stream.</p>
<h1 class="instructions">Description</h1>
<p class="instructions">This test checks that MediaStream.getTrackById behaves as expected</p>

<script>

promise_test(async t => {
  const stream = await navigator.mediaDevices.getUserMedia({video: true});
  const track = stream.getVideoTracks()[0];
  t.add_cleanup(() => track.stop());
  assert_equals(track, stream.getTrackById(track.id), "getTrackById returns track of given id");
  assert_equals(stream.getTrackById(track.id + "foo"), null, "getTrackById of inexistant id returns null");
}, "Tests that MediaStream.getTrackById works as expected");

</script>
